+++ src/EVM.hs
@@ -14,6 +14,7 @@ import EVM.ABI
 import EVM.Expr (readStorage, concStoreContains, writeStorage, readByte, readWord, writeWord,
   writeByte, bufLength, indexWord, readBytes, copySlice, wordToAddr, maybeLitByteSimp, maybeLitWordSimp, maybeLitAddrSimp)
 import EVM.Expr qualified as Expr
+import EVM.HashCons qualified as HashCons
 import EVM.FeeSchedule (FeeSchedule (..))
 import EVM.FeeSchedule qualified as Fees (feeSchedule)
 import EVM.Merge qualified as Merge
@@ -497,7 +498,7 @@ exec1 conf = do
                           modifying #keccakPreImgs (insert (bs, kc))
                           pure $ Lit kc
                         )
-                    buf -> pure $ Keccak buf
+                    buf -> pure $ HashCons.internExpr (Keccak buf)
                   next
                   assign' (#state % #stack) (hash : xs)
             _ -> underrun
@@ -795,7 +796,8 @@ exec1 conf = do
 
                 symbolicRead :: EVM t () = if this.external
                   then accessStorage self x finalizeLoad
-                  else finalizeLoad $ Expr.readStorage' (Expr.concKeccakOnePass x) this.storage
+                  else let slot = HashCons.internExpr (Expr.concKeccakOnePass x)
+                       in finalizeLoad $ HashCons.internExpr (Expr.readStorage' slot this.storage)
 
                 concreteRead :: EVM t () = do
                   acc <- accessStorageForGas self (forceLit x)
@@ -3277,7 +3279,7 @@ stackOp1 cost f =
     x:xs ->
       burn cost $ do
         next
-        let !y = f x
+        let !y = HashCons.internExpr (f x)
         assign' (#state % #stack) $ y : xs
     _ ->
       underrun
@@ -3292,7 +3294,7 @@ stackOp2 cost f =
     x:y:xs ->
       burn cost $ do
         next
-        assign' (#state % #stack) $ f x y : xs
+        assign' (#state % #stack) $ HashCons.internExpr (f x y) : xs
     _ ->
       underrun
 
@@ -3306,7 +3308,7 @@ stackOp3 cost f =
     x:y:z:xs ->
       burn cost $ do
       next
-      assign' (#state % #stack) $ f x y z : xs
+      assign' (#state % #stack) $ HashCons.internExpr (f x y z) : xs
     _ ->
       underrun